requirements.txt.txt
pdfplumber
sentence-transformers
faiss-cpu
gradio
numpy